HAN: add: news 기능 결합 및 세부기능 수정
