+++ package.json
@@ -6,6 +6,7 @@
     "@emotion/react": "^11.11.1",
     "@emotion/styled": "^11.11.0",
     "@mui/material": "^5.14.5",
+    "@mui/styled-engine": "^5.13.2",
     "@mui/styled-engine-sc": "^5.12.0",
     "@redux-devtools/extension": "^3.2.5",
     "@testing-library/jest-dom": "^5.17.0",
@@ -25,7 +26,7 @@
     "react-js-pagination": "^3.0.3",
     "react-paginate": "^8.2.0",
     "react-redux": "^8.1.2",
-    "react-router-dom": "^6.14.2",
+    "react-router-dom": "^6.15.0",
     "react-scripts": "5.0.1",
     "redux": "^4.2.1",
     "redux-actions": "^3.0.0",
