+++ src/App.js
@@ -1,15 +1,12 @@
-import {
-  Routes,
-  Route,
-  useNavigationType,
-  useLocation,
-} from "react-router-dom";
+import { Routes, Route, useNavigationType, useLocation } from "react-router-dom";
 import { useEffect } from "react";
 
 import Home from "./pages/Home";
 import LoginPage from "./pages/LoginPage";
 import RegisterPage from "./pages/RegisterPage";
 import News from "./pages/News";
+import Header from "./components/common/Header";
+import Footer from "./components/common/Footer";
 
 function App() {
   const action = useNavigationType();
@@ -49,21 +46,19 @@ function App() {
   }, [pathname]);
 
   return (
-    <Routes>
-      <Route path="/" element={<Home />} />
-      <Route
-        path="/login"
-        element={
-          <>
-            <Home />
-            <LoginPage />
-          </>
-        }
-      />
-      <Route path="/register" element={<RegisterPage />} />
-      <Route path="/news/list" element={<News />} />
-      <Route path="/news/list/:page" element={<News />} />
-    </Routes>
+    <div>
+      <Header />
+      <Routes>
+        <Route path="/" element={<Home />} />
+        <Route path="/login" element={<><Home /><LoginPage /></>} />
+        <Route path="/register" element={<RegisterPage />} />
+        <Route path="/news/list" element={<News />} />
+        <Route path="/news/list/:pageNum" element={<News />} />
+        <Route path="/news/search/:keyword" element={<News />} />
+        <Route path="/news/search/:keyword/:pageNum" element={<News />} />
+      </Routes>
+      <Footer />
+    </div>
   );
 }
 export default App;
