--- src/components/UpperBanner.js
@@ -1,19 +1,0 @@
-import styles from "../styles/UpperBanner.module.css";
-
-const UpperBanner = () => {
-  return (
-    <div className={styles.divcss1mvkryr}>
-      <div className={styles.linkPc} />
-      <div className={styles.linkPc1} />
-      <div className={styles.divcss1rgilik}>
-        <img className={styles.linkPc2} alt="" src="/link--pc-@2x.png" />
-        <img className={styles.linkPc2} alt="" src="/link--pc-1@2x.png" />
-      </div>
-      <div className={styles.button}>
-        <img className={styles.vectorIcon} alt="" src="/vector18.svg" />
-      </div>
-    </div>
-  );
-};
-
-export default UpperBanner;
