+++ src/pages/News.js
@@ -1,18 +1,16 @@
 import React from "react";
 
-import Header from "../components/Header";
-import NewsList from "../components/NewsList";
+import NewsList from "../components/news/NewsList";
 
-import styles from "../styles/News.module.css";
+import styles from "../styles/pages/News.module.css";
 
 const News = () => {
 
   return (
     <div className={styles.newsContainer}>
-      <Header />
       <div className={styles.newsListContainer}>
         <NewsList />
-        </div>        
+      </div> 
     </div>
   );
 };
