--- src/styles/UpperBanner.module.css
@@ -1,65 +1,0 @@
-.linkPc,
-.linkPc1 {
-  position: absolute;
-  height: 100%;
-  width: 50%;
-  top: 0;
-  right: 50%;
-  bottom: 0;
-  left: 0;
-  background-color: var(--color-black);
-}
-.linkPc1 {
-  right: 0;
-  left: 50%;
-  background-color: var(--color-deepskyblue-200);
-}
-.linkPc2 {
-  align-self: stretch;
-  flex: 1;
-  position: relative;
-  max-width: 100%;
-  overflow: hidden;
-  max-height: 100%;
-  object-fit: cover;
-}
-.divcss1rgilik {
-  position: absolute;
-  top: 0;
-  left: calc(50% - 638px);
-  width: 1156px;
-  display: flex;
-  flex-direction: row;
-  padding: var(--padding-0) var(--padding-41xl);
-  box-sizing: border-box;
-  align-items: flex-start;
-  justify-content: flex-start;
-  max-width: 1156px;
-}
-.vectorIcon {
-  position: relative;
-  width: 17.44px;
-  height: 17.44px;
-}
-.button {
-  position: absolute;
-  top: 36px;
-  left: 1876px;
-  width: 24px;
-  height: 25px;
-  overflow: hidden;
-  display: flex;
-  flex-direction: row;
-  padding: 3.78125px 6.5625px 3.78125px var(--padding-0);
-  box-sizing: border-box;
-  align-items: flex-start;
-  justify-content: flex-start;
-}
-.divcss1mvkryr {
-  position: absolute;
-  width: 100%;
-  top: 0;
-  right: 0;
-  left: 0;
-  height: 50px;
-}
